Added a new exploreReadExperimental page.
Moved exploreRead and exploreReadExperimental to AssemblerHttpServer-Reads.cpp. Once the new
page is
up to snuff, I plan to drop the old one.

Summary:
1. Old Reads page is unchanged.
2. There's a new Reads(Experimental) page which has all the changes.
3. The new page shows the selected* portion of the run-length sequence, with some caveats.

(*) Selection still happens using raw sequence positions. I map those to positions in RLE. The
start
of the RLE display happens at the largest non-negative multiple of 10 <= the mapped begin RLE
position. This makes existing labeling work as is with a new start and end value. Only those
k-mers
which map entirely into the range are shown on the page.